Build on every commit
