+++ app/NX/NXAdmin/index.tsx
@@ -1,5 +1,7 @@
 import NXAdmin from './NXAdmin';
+import NXAdminBtn from './components/NXAdminBtn';
 
 export {
     NXAdmin,
+    NXAdminBtn,
 };
