+++ app/NX/Virus/Virus.tsx
@@ -9,7 +9,7 @@ import {
 } from 'react-share';
 import { Box, Typography, ButtonBase, Popover } from '@mui/material';
 import { Icon } from '../../NX/DesignSystem';
-import { Forward } from '../Virus'
+// import { Forward } from '../Virus'
 
 export default function Virus({ meta }: { meta: T_Meta }) {
   if (!meta) return <>
@@ -35,9 +35,9 @@ export default function Virus({ meta }: { meta: T_Meta }) {
       gap: 2,
       mt: 4
     }}>
-      <Box>
+      {/* <Box>
         <Forward />
-      </Box>
+      </Box> */}
       <Box>
         <ButtonBase
           onClick={e => {
